Move lock types into wire-types
Move the Lock types into wire-types so they can be shared between the
browser and server.
