+++ iml-agent-comms/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2018"
 [dependencies]
 env_logger = "0.6.1"
 failure = "0.1.5"
-futures = "0.1.27"
-iml-wire-types = { path = "../iml-wire-types", version = "0.1.4" }
+futures = "0.1.28"
+iml-wire-types = { path = "../iml-wire-types", version = "0.1" }
 iml-rabbit = { path = "../iml-rabbit", version = "0.1.0" }
 iml-manager-env = { path = "../iml-manager-env", version = "0.1.0" }
 lapin-futures = "0.18.0"
@@ -17,6 +17,6 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.39"
 tokio = "0.1"
 uuid = { version = "0.7", features = ["v4"] }
-warp = "0.1.15"
+warp = "0.1.16"
 parking_lot = "0.8"
