+++ iml-agent/Cargo.toml
@@ -14,7 +14,7 @@ csv = "1"
 dns-lookup = "1"
 env_logger = "0.6.1"
 exitcode = "1.1"
-futures = "0.1.27"
+futures = "0.1.28"
 lazy_static = "1.3.0"
 termion = "1"
 spinners = "1.0.0"
@@ -26,10 +26,10 @@ hyper = "0.12"
 http = "0.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-structopt = "0.2.16"
+structopt = "0.2.18"
 tokio = "0.1"
 tokio-timer = "0.2.11"
-tokio-process = "0.2.3"
+tokio-process = "0.2.4"
 tokio-threadpool = "0.1"
 tokio-tls = "0.2"
 native-tls = "0.2"
@@ -37,7 +37,7 @@ url = "1.7.2"
 v_hist = "0.1.1"
 uuid = { version = "0.7", features = ["v4"] }
 parking_lot = "0.8"
-iml-wire-types = { path = "../iml-wire-types", version = "0.1.4" }
+iml-wire-types = { path = "../iml-wire-types", version = "0.1" }
 liblustreapi = { path = "../liblustreapi", version = "0.1" }
 iml-fs = { path = "../iml-fs", version = "0.1.0" }
 stream-lines = { git = "https://github.com/softprops/stream-lines" }
