+++ iml-manager-cli/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 dotenv = "0.14"
 env_logger = "0.6.1"
 futures = "0.1"
-iml-wire-types = { path = "../iml-wire-types", version = "0.1.4" }
+iml-wire-types = { path = "../iml-wire-types", version = "0.1" }
 iml-manager-client = { path = "../iml-manager-client", version = "0.1.0" }
 log = "0.4.6"
 prettytable-rs = "0.8"
