+++ iml-services/Cargo.toml
@@ -6,16 +6,16 @@ edition = "2018"
 
 [dependencies]
 lapin-futures = "0.18.0"
-futures = "0.1.26"
-tokio = "0.1.19"
-tokio-timer = "0.2.10"
+futures = "0.1.28"
+tokio = "0.1.21"
+tokio-timer = "0.2.11"
 env_logger = "0.6.1"
 log = "0.4.6"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 failure = "0.1.5"
-warp = "0.1.15" 
-iml-wire-types = { path = "../iml-wire-types", version = "0.1.4" }
+warp = "0.1.16" 
+iml-wire-types = { path = "../iml-wire-types", version = "0.1" }
 iml-rabbit = { path = "../iml-rabbit", version = "0.1.0" }
 iml-manager-env = { path = "../iml-manager-env", version = "0.1.0" }
 tokio-runtime-shutdown = { path = "../tokio-runtime-shutdown", version = "0.1.0" }
