+++ iml-wasm-components/iml-action-dropdown/Cargo.toml
@@ -16,7 +16,7 @@ cfg-if = "0.1"
 log = "0.4"
 iml-tooltip = { path = "../iml-tooltip", version = "0.1.0" }
 iml-sleep = { path = "../iml-sleep", version = "0.1.0" }
-iml-wire-types = "0.1.2"
+iml-wire-types = "0.1.5"
 console_log = { version = "0.1", optional = true }
