+++ iml-wasm-components/iml-alert-indicator/Cargo.toml
@@ -8,5 +8,5 @@ edition = "2018"
 seed = "=0.3.7"
 log = "0.4"
 iml-tooltip = { path = "../iml-tooltip", version = "0.1.0" }
-iml-wire-types = "0.1.4"
+iml-wire-types = "0.1.5"
 iml-popover = { path = "../iml-popover", version = "0.1" }
