+++ iml-wire-types/Cargo.toml
@@ -2,7 +2,7 @@
 name = "iml-wire-types"
 description = "Shared types for the IML project"
 license = "MIT"
-version = "0.1.4"
+version = "0.1.5"
 authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
